# Conflicts:
#	.github/.gitignore
#	.github/build.sh
#	changelog.txt
#	data/IntroText.h
#	resources/icon.ico
#	resources/icon/powder-128.png
#	resources/icon/powder-16.png
#	resources/icon/powder-24.png
#	resources/icon/powder-256.png
#	resources/icon/powder-32.png
#	resources/icon/powder-48.png
#	src/simulation/elements/WARP.cpp
#	src/simulation/elements/meson.build

+++ src/simulation/elements/CEXP.cpp
@@ -76,7 +76,7 @@ static int update(UPDATE_FUNC_ARGS)
 				}
 				if (parts[i].tmp2 > 0)
 				{
-					sim->pv[(y / CELL) + ry][(x / CELL) + rx] = parts[i].life;
+					sim->pv[(y / CELL) + ry][(x / CELL) + rx] = (float)(parts[i].life);
 					sim->create_part(-1, x + parts[i].tmp, y + parts[i].tmp, PT_PLSM);
 					sim->create_part(-1, x - parts[i].tmp, y - parts[i].tmp, PT_PLSM);
 					sim->create_part(-1, x - parts[i].tmp, y + parts[i].tmp, PT_PLSM);
@@ -91,5 +91,5 @@ static void create(ELEMENT_CREATE_FUNC_ARGS)
 {
 	sim->parts[i].temp = 9720;
 	sim->parts[i].tmp = 10;
-	sim->parts[i].life = 256.0;
+	sim->parts[i].life = 256;
 }
